requirements.txt
streamlit>=1.25.0
PyMuPDF>=1.23.0
chromadb>=0.4.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.1.0
langchain>=0.1.0
langgraph>=0.0.26
python-dotenv>=1.0.0
langsmith>=0.4.0 